## Healthcare-AI-chatbot
The healthcare AI chatbot is a project of absolute perfection.

The field of medicals and healthcare is a field that uses a lot of data, so the use of algorithms and AI is pretty simple and logical.

This project is a bot that has multiple functionalities and a concise algorithm.

This project is a simple healthcare chatbot built using Python. The chatbot uses natural language processing (NLP) to understand user queries, identify intents, extract relevant entities, and generate appropriate responses.
It aims to build an interactive chatbot that can assist users with healthcare-related inquiries.

it is all presented to you by me and a fellow programmer ashu-devx.
don't forget to check out his profile.
https://github.com/ashu-devx

## Prerequisites

- **Python 3.x**: Make sure Python is installed on your system.
- **Virtual Environment**: It's recommended to use a virtual environment to manage project dependencies.


## Use cases
    researching medical solutions
    
    

## Functionalities
    Using NLU
    Using knowledge base integration
    An intercative user interface
    ensuring accurate response generation

## Installation

1. **Clone the repository**:
    ```bash
    git clone https://github.com/your-username/healthcare_chatbot.git
    cd healthcare_chatbot
    ```

2. **Create and activate a virtual environment**:
    ```bash
    python -m venv chatbot_env
    source chatbot_env/bin/activate  # On Windows use `chatbot_env\Scripts\activate`
    ```

3. **Install necessary libraries**:
    ```bash
    pip install -r requirements.txt
    python -m spacy download en_core_web_sm
    ```

## Project Structure